use snake_case and move extract_* functions
extract_value function was changed to use a regular expression which will make it easier to use it as an Adapter level preprocess_value function as proposed in #331
